+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Doesn't overwrite `Content-Type` if already set.
+- Deletes the `json` property before calling `fetch`.
 
 ## 0.1.337 (Dec 11, 2015)
 - Scoped release for personal use and profit.